feat: implement user authorization checks in task use cases and add NotAuthorizedToPerformError

+++ src/application/use-cases/task/DeleteTaskUseCase.ts
@@ -1,13 +1,16 @@
-import { TaskNotFoundError } from '../../errors/TaskNotFoundError.js'
 import type { TaskRepository } from '../../../domain/repositories/TaskRepository.js'
 import type { ID, SuccessMessage } from '../../../domain/types/types.js'
+import { NotAuthorizedToPerformError } from '../../errors/NotAuthorizedToPerformError.js'
+import { TaskNotFoundError } from '../../errors/TaskNotFoundError.js'
 
 export class DeleteTaskUseCase {
   constructor(private readonly taskRepository: TaskRepository) {}
 
-  async execute(id: ID): Promise<SuccessMessage> {
+  async execute(id: ID, requestedById: ID): Promise<SuccessMessage> {
     const task = await this.taskRepository.findById(id)
     if (task == null) throw new TaskNotFoundError()
+
+    if (task.userId !== requestedById) throw new NotAuthorizedToPerformError()
 
     const successMessage = await this.taskRepository.deleteById(id)
     return successMessage
